src/liveapi/requirements.txt
alembic==1.0.0
aniso8601==3.0.2
asn1crypto==0.24.0
atomicwrites==1.2.1
attrs==18.2.0
bcrypt==3.1.4
cachetools==2.1.0
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
Click==7.0
cryptography==2.3.1
enum-compat==0.0.2
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Caching==1.4.0
Flask-Cors==3.0.6
Flask-JWT-Extended==3.13.1
Flask-Migrate==2.2.1
flask-restplus==0.12.1
Flask-Script==2.0.6
Flask-Session==0.3.1
Flask-SQLAlchemy==2.3.2
Flask-Testing==0.7.1
ForgeryPy3==0.3.1
google-api-python-client==1.7.4
google-auth==1.5.1
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.2.0
gunicorn==19.9.0
httplib2==0.11.3
idna==2.7
itsdangerous==0.24
Jinja2==2.10
jsonschema==2.6.0
Mako==1.0.7
MarkupSafe==1.0
mock==2.0.0
more-itertools==4.3.0
oauth2==1.9.0.post1
oauthlib==2.1.0
pbr==4.3.0
pluggy==0.7.1
psycopg2==2.7.5
py==1.6.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.19
PyJWT==1.6.4
pyOpenSSL==18.0.0
pytest==3.8.2
python-box==3.2.0
python-dateutil==2.7.3
python-dotenv==0.9.1
python-editor==1.0.3
pytz==2018.5
redis==2.10.6
requests==2.19.1
requests-oauthlib==1.0.0
rsa==4.0
six==1.11.0
SQLAlchemy==1.2.12
SQLAlchemy-Utils==0.33.5
uritemplate==3.0.0
urllib3==1.23
Werkzeug==0.14.1
